MyDaggerFragmentでLayoutIdを受け取る
