+++ app/src/main/java/com/shebang/dog/goo/ui/tab/MyDaggerFragment.kt
@@ -1,0 +1,18 @@
+package com.shebang.dog.goo.ui.tab
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import dagger.android.support.DaggerFragment
+
+abstract class MyDaggerFragment(@LayoutRes private val layoutResId: Int) : DaggerFragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(layoutResId, container, false)
+    }
+}
