+++ app/src/main/java/com/shebang/dog/goo/ui/tab/TabbedFragment.kt
@@ -1,24 +1,13 @@
 package com.shebang.dog.goo.ui.tab
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.annotation.DrawableRes
 import androidx.annotation.LayoutRes
-import dagger.android.support.DaggerFragment
 
-abstract class TabbedFragment(@LayoutRes private val layoutResId: Int) : DaggerFragment() {
+abstract class TabbedFragment(@LayoutRes private val layoutResId: Int) :
+    MyDaggerFragment(layoutResId) {
+
     @DrawableRes
     abstract fun getTabIconId(): Int
 
     abstract fun getTabTitle(): String
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(layoutResId, container, false)
-    }
 }
